refactor(frontend): split detailed usage page
